classes/PasswordHelper.py
import bcrypt

class PasswordHelper:
    @staticmethod
    # Get the hash and salt for a plain text pw
    def get_hashed_password(plain_text_password):
        # Hash a password for the first time
        #   (Using bcrypt, the salt is saved into the hash itself)
        return bcrypt.hashpw(plain_text_password, bcrypt.gensalt())

    @staticmethod
    # Check a plain text user pw vs a hashed version. 
    # Returns bool i think?
    def check_password(plain_text_password, hashed_password):
        # Check hashed password. Using bcrypt, the salt is saved into the hash itself
        return bcrypt.checkpw(plain_text_password, hashed_password)

    