requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==3.0.1
MarkupSafe==2.1.2
psycopg2==2.9.6
py-bcrypt==0.4
python-dotenv==1.0.0
Werkzeug==2.3.4
zipp==3.15.0
